refactor: move getBucketStats/clear to test-only subclass
